move exceptions defination to exceptions.h
